[FS-191] fix: component Item
