+++ src/stories/components/SocialNetworks/components/Item/styles.ts
@@ -1,0 +1,8 @@
+import { Typography } from '@mui/material';
+import { styled } from '@mui/material/styles';
+import { spacing } from 'theme/spacing';
+
+export const IconName = styled(Typography)`
+  display: flex;
+  gap: ${spacing.S10};
+`;
