+++ src/stories/components/SocialNetworks/components/Item/types.ts
@@ -1,0 +1,5 @@
+export interface ItemProps {
+  id: string;
+  name: string;
+  redirectLink: string;
+}
